[feat] 버블차트 컨테이너 레이아웃 작업
