+++ frontend/components/community/index.ts
@@ -2,6 +2,7 @@ import AutoComplete from './AutoComplete';
 import CommunityHeader from './CommunityHeader';
 import CommunityLayout from './CommunityLayout';
 import KeywordAddModal from './KeywordAddModal';
+import KeywordBubbleChart from './KeywordBubbleChart';
 import MyKeyword from './MyKeyword';
 import MyKeywordList from './MyKeywordList';
 
@@ -10,6 +11,7 @@ export {
   CommunityHeader,
   CommunityLayout,
   KeywordAddModal,
+  KeywordBubbleChart,
   MyKeyword,
   MyKeywordList,
 };
