+++ frontend/pages/community/[id].tsx
@@ -5,11 +5,16 @@ import {
   CommunityHeader,
   CommunityLayout,
   KeywordAddModal,
+  KeywordBubbleChart,
 } from '@components/community';
 import { useEffect, useState } from 'react';
+import { KeywordData } from '../../types/types';
 
 const Community = () => {
   const [userData, setUserData] = useState<string | null>('');
+  const [communityKeywordData, setCommunityKeywordData] = useState<
+    KeywordData[]
+  >([]);
   const [isOpenLoginModal, setIsOpenLoginModal] = useState<boolean>(false);
   const [isOpenKeywordModal, setIsOpenKeywordModal] = useState<boolean>(false);
 
@@ -22,8 +27,21 @@ const Community = () => {
   };
 
   useEffect(() => {
+    // TODO : 나중에는 서버에서 받아와야한다.
+    const mockWordList = [
+      { keyword: '가', count: 3 },
+      { keyword: '다', count: 3 },
+      { keyword: '나', count: 3 },
+      { keyword: '가나', count: 3 },
+      { keyword: '가다', count: 3 },
+      { keyword: '가나다', count: 2 },
+      { keyword: '가나라', count: 3 },
+      { keyword: '가나라마', count: 3 },
+    ];
+
     const username = localStorage.getItem('waglewagle-username');
     setUserData(username);
+    setCommunityKeywordData(mockWordList);
   }, []);
 
   return (
@@ -34,7 +52,8 @@ const Community = () => {
         handleClickKeywordModal={handleClickKeywordModal}
         handleClickEnter={handleClickEnter}
       />
-      <AutoComplete />
+      <KeywordBubbleChart communityKeywordData={communityKeywordData} />
+      <AutoComplete communityKeywordData={communityKeywordData} />
       <Modal
         isOpenModal={isOpenLoginModal}
         closeModal={() => setIsOpenLoginModal(false)}
