+++ frontend/sass/components/community/KeywordBubbleChart.module.scss
@@ -1,0 +1,3 @@
+.chart-container {
+  display: flex;
+}
