+++ frontend/types/types.ts
@@ -1,0 +1,4 @@
+export type KeywordData = {
+  keyword: string;
+  count: number;
+};
